Refactor to split components
